requirements.txt
bs4
langchain
openai
chromadb
pypdf
python-dotenv
tiktoken
faiss-cpu